move html simplification into kalosm language
